Add an idle timer for refreshing the speedbar.
This is different than speedbar-auto-refresh, because the auto refresh
has the annoying property of changing the speedbar to only display the
immediate folder that my current buffer lives in, often hiding many
other folders that I want to keep in view. As coded, this causes the
speedbar to simply check for the presence of new files, without
changing which directory is displayed.

The end result is that I can create new files outside of emacs, and
the speedbar will update to show them after a reasonably short delay.